+++ examples/miniapps/application-multiple-containers/example/__main__.py
@@ -2,23 +2,29 @@
 
 import sys
 
+from dependency_injector.wiring import Provide
+
+from .services import UserService, AuthService, PhotoService
 from .containers import Application
 
 
-def main(email: str, password: str, photo: str) -> None:
-    application = Application()
-
-    application.config.from_yaml('config.yml')
-    application.core.configure_logging()
-
-    user_service = application.services.user()
-    auth_service = application.services.auth()
-    photo_service = application.services.photo()
-
+def main(
+        email: str,
+        password: str,
+        photo: str,
+        user_service: UserService = Provide[Application.services.user],
+        auth_service: AuthService = Provide[Application.services.auth],
+        photo_service: PhotoService = Provide[Application.services.photo],
+) -> None:
     user = user_service.get_user(email)
     auth_service.authenticate(user, password)
     photo_service.upload_photo(user, photo)
 
 
 if __name__ == '__main__':
+    application = Application()
+    application.config.from_yaml('config.yml')
+    application.core.configure_logging()
+    application.wire(modules=[sys.modules[__name__]])
+
     main(*sys.argv[1:])