+++ examples/miniapps/django/githubnavigator/__init__.py
@@ -1,0 +1,8 @@
+"""Project package."""
+
+from .containers import Container
+from . import settings
+
+
+container = Container()
+container.config.from_dict(settings.__dict__)