+++ examples/miniapps/movie-lister/movies/tests.py
@@ -4,12 +4,12 @@ from unittest import mock
 
 import pytest
 
-from .containers import ApplicationContainer
+from .containers import Container
 
 
 @pytest.fixture
 def container():
-    container = ApplicationContainer()
+    container = Container()
     container.config.from_dict({
         'finder': {
             'type': 'csv',