+++ docs/main/changelog.rst
@@ -7,6 +7,44 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+4.0.0
+-----
+New features:
+
+- Add ``wiring`` feature.
+
+Deprecations:
+
+- Deprecate ``ext.aiohttp`` module in favor of ``wiring`` feature.
+- Deprecate ``ext.flask`` module in favor of ``wiring`` feature.
+- Deprecate ``.delegate()`` provider method in favor of ``.provider`` attribute.
+
+Removals:
+
+- Remove deprecated ``types`` module.
+
+Tutorials:
+
+-  Update ``flask`` tutorial.
+-  Update ``aiohttp`` tutorial.
+-  Update ``asyncio`` daemon tutorial.
+-  Update CLI application tutorial.
+
+Examples:
+
+- Add ``django`` example.
+- Add ``sanic`` example.
+- Update ``aiohttp`` example.
+- Update ``flask`` example.
+- Update ``asyncio`` daemon example.
+- Update ``movie-lister`` example.
+- Update CLI application example.
+
+Misc:
+
+- Regenerate C sources using Cython 0.29.21.
+- Improve documentation and README (typos removal, rewording, etc).
+
 3.44.0
 ------
 - Add native support of the generics to the providers: ``some_provider = providers.Provider[SomeClass]``.